fix searchbar tests
