+++ lib/components/SearchBarAndListings.js
@@ -1,6 +1,7 @@
 import { Navbar, FormGroup, FormControl, Button } from 'react-bootstrap';
 import React from 'react';
 import JobListings from './JobListings';
+var $ = require('jquery');
 
 class SearchBarAndListings extends React.Component {
   constructor() {
